puffin: Add ExtentStream
We need some sort of extent stream for the update payload tests in
update_engine. These scripts are run as a part of canary builds. This CL
adds a new class ExtentStream that is only used in main.cc for applying
patches. This class should not be shipped to the client devices as part
of puffpatch nor it needs to be in puffdiff.

BUG=chromium:768461
TEST=unittest pass;

Change-Id: Ic482be8e9b6b7e7f12f5f796de1d96d57275097b
Reviewed-on: https://chromium-review.googlesource.com/714453
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Sen Jiang <senj@chromium.org>