puffin: Remove primitive caching for puffpatch
Currently we read the entire source into a buffer in puffpatch. This is
problematic memory-wise if the source files are large (which they are). It is
better to not do this for now to allow puffing to be done without any caching
until a better caching scheme is added.

BUG=chromium:775072
TEST=unittests passed;

Change-Id: I5a47147fdc1cf5a76cca6700cb16f9cae572cee7
Reviewed-on: https://chromium-review.googlesource.com/722451
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Sen Jiang <senj@chromium.org>