puffin: Cache puff stream in puffpatch
Currently without caching the puff stream, the OS updates can be very slow. For
example, a back of envelope calculation can show that if we don't cache the puff
stream, we may need to read and puff the entire deflate stream (about 500 MB)
160 times which can be very slow. And examination shows that updates can take
more than one hour to finish on a veyron_minnie. It may not be reasonable to
keep memory usage low but keep the CPU usage high for extended periods of
time. The other rationale behind caching the puff stream is that if we only use
bsdiff, then we have to keep large patches in memory when applying the
patch. But since puffin reduced the patch size by a large factor (so to keep a
much smaller patch in memory), then it won't be as bad to cache some memory for
faster patching. It is basically a trade-off between CPU time and memory. We may
end up using the same amount of memory (which in practice is not going to
happen, we will use much less), but the payload patch size is much smaller.

This patch modifies PuffinStream to cache only the Puff Buffers up to a maximum
passed total memory size as a parameter to puffpatch. This acts as an LRU, the
least recently used puff buffers are ejected when the cache is full.

BUG=chromium:775072
TEST=unittests passed; brillo_update_payload verfiy passes; cros flash with delta updates passes;

Change-Id: I2a390eb35eb5449577cb5e2b1e857bcdca093f68
Reviewed-on: https://chromium-review.googlesource.com/722452
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>