Fix a bug in huffman table for huffing
When we are creating the huffman table contents for huffing, there is a point
where we do not check for the access boundary of the vector codeindexpair_. This
causes possible invalid symbols be inserted into rcodes vector. But this only
causes problem if the input puff stream is invalid. If the input puff stream is
valid, those invalid symbols will never get touched. Also the out of boundary
will not cause heap access issues because we have always pre allocated a larger
buffer. Another issue was that we were only allocating 286 bytes for
codeindexpairs_, but for fixed huffman codes 288 bytes needed. We increase the
pre-allocation to prevent a realloc.

This bug was found through asan sanitizer builders.

Bug: 72708841
Test: unittests pass.

Change-Id: Ia2a55a4048e1d40746ee831e556877f8a53d7e3f