Allow parsing compressed files in puffin
This patch adds a new flag src_file_type the puffin executable to allow
parsing an input file without the need to manually give the deflate
locations. We can later add dst_file_type to do puffpatch between
different types of files, but it is not necessary at this point yet.

Bug: 73487244
Test: puffin with src_file_type of zlib, deflate and zip passed.

Change-Id: I1addc8f1567bda652c4f138f0e9f825eb7ece4f8