Start detecting file types
This patch adds option for detecting file types based on the extension of the
given file paths. However this file type can be overriden by explicitly passing
src_file_type or dst_file_type. Btw, In turns out it is a good idea to add
dst_file_type flag to so we can better test the patching mechanism.

Bug: 73487244
Test: puffin --verbose --operation=puffdiff --src_file_type=zip --src_file=zipme.zip --dst_file=big.txt.1.gz --dst_file_type=gzip --patch_file=patch.bin

Change-Id: Ib860b68acc822a256bd56cdda30dc59273b878c2