Convert majority of size_t to uint64_t for 32-bit compatibility
We need to get puffin to work correctly on 32-bit machines. This patch converts
appropriate size_t types to uint64_t to make sure when we are working with large
files in 32bit machines, nothing breaks.

The 32 bit tests actually found some errors. For example BitReader and BitWriter
should have their size and indices kept in 64 bit integers otherwise a multiply
by 8 (to figure out the bit-based length) will overflow. Although it has not
been tested for puffdiff and puffpatch.

Bug: 73781483
Test: unittests pass
Test: puffhuff on a >5G zlib file passes
Test: build with 32-bit compiler, and puffhuff on >1GB passes

Change-Id: I0234a787e095a7ebf0489966a7b7565b6688496c