Add missing invalid distance check
In the puffin spec, we defined the distance on the wire to be zero
based ([1..32768] on deflate RFC). But I missed it in the implementation and
stored non-zero based distances. However, this is not a bug and since we have
not shipped it yet, it is possible to change.

Another change include, adding proper checks for finding invalid copy length or
distance. Also more unittests for invalid copy length/distance values are
added. Lack of these checks were found by running puffin_fuzzer on cluster fuzz.

Bug: crbug.com/817733
Bug: crbug.com/817686
Test: unittests
Test: test_corpus.py

Change-Id: I38bf630904d7996a3c4f15919960517d26520987