Remove anything related to Errors
In the past we used to check for errors to make runtime decision based on the
error code. For example, we retrying the puff operation if the output buffer was
not enought. But puffin does not function that way anymore. So it is fine to
remove all error related code. This will slightly reduce the overhead both on
the code size and performance.

Bug: 74992491
Test: unittests
Change-Id: I5fd0353189be5abda75725475b1b5986eab87e68