Prevent BsdiffStream from accepting a nullptr input sream
The BsdiffStream defined in puffpatch.cc can accept a nullptr stream with no
problem. This is problematic. This patch adds a factory function to
BsdiffFileStream to check if the incoming stream pointer is invalid, in which
case it will return nullptr. Also we check that the source and target streams
are created correctly.

This bug was found by the puffin_fuzzer

Bug: crbug.com/830156
Bug: crbug.com/831772
Test: unittests
Test: corpus test

Change-Id: Iad8b093166ef26e2649e14e7e78adee17a5b5557